fix: 사용자 activity 저장 1. activity entity 필드 변경

+++ user-service/src/main/java/com/example/user/repository/FollowRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface FollowRepository extends JpaRepository<Follow, Integer> {
+public interface FollowRepository extends JpaRepository<Follow, Long> {
     Optional<Follow> findByFollowerIdAndFollowingId(String followerId, String followingId);
 }
